git-svn-id: https://mobicents.googlecode.com/svn/trunk/protocols/ss7@12671 bf0df8d0-2c1f-0410-b170-bd30377b63dc

+++ isup/isup-api/src/main/java/org/mobicents/protocols/ss7/isup/message/parameter/CircuitIdentificationCode.java
@@ -14,7 +14,7 @@ package org.mobicents.protocols.ss7.isup.message.parameter;
  */
 public interface CircuitIdentificationCode extends ISUPParameter {
 
-	public long getCIC();
-	public void setCIC(long cic);
+	public int getCIC();
+	public void setCIC(int cic);
 	
 }
